+++ lib/bot.js
@@ -1,6 +1,6 @@
 var
   parser = require("./parser"),
-  command = require("./commands"),
+  command = require("./process"),
   Promise = require("BlueBird"),
   sendMail = require("./sendMail"),
   report = require("./report");
