+++ lib/commands.js
@@ -1,43 +1,21 @@
-/**
- * Various command, used by bot
- */
+var command = [{
+  name: "!help",
+  man: "To show any help"
+}, {
+  name: "addt",
+  man: "add transaction"
+}, {
+  name: "addm",
+  man: "add memebers"
+}, {
+  name: "export xls <month>",
+  man: "export xls for <month>"
+}, {
+  name: "mail <user-name> or <email-id> for <month>",
+  man: "mail user or email"
+}, {
+  name: "show report of <user-name> for <month>",
+  man: "show report of user for month"
+}]
 
-var
-  mongoose = require('mongoose'),
-  Expense = mongoose.model('Expense'),
-  Member = mongoose.model('Member'),
-  _ = require('underscore'),
-  Promise = require("bluebird");
-
-module.exports = {
-  record: function(from, transaction, message) {
-    var query = {};
-    if (transaction.userName !== "all") {
-      query.name = transaction.userName;
-    }
-
-    return new Promise(function(resolve, reject) {
-      Member.find(query, function(err, members) {
-        var paidBy = _(members).find(function(member) {
-          return member.name === from;
-        });
-        var paidFor = _(members).chain().filter(function(member) {
-          return member.name !== from;
-        }).pluck("_id").value();
-
-        var share = transaction.amount / members.length;
-
-        Expense.credit(share, paidBy, paidFor, message, transaction.date)
-          .then(function(result) {
-            resolve("transaction recorded.");
-          })
-          .catch(function(err) {
-            reject(err);
-          });
-      });
-    });
-  },
-  addMember: Promise.method(function(name) {
-    return Member.addMember(name);
-  })
-}
+module.exports = command
